Move tests to separate module
